pull drag interaction out of shapes_test - and speed it up
